+++ lib/mc.c
@@ -1968,7 +1968,7 @@ startup_got_sel_time(ipmi_mc_t  *mc,
     gettimeofday(&now, NULL);
     time = ipmi_get_uint32(rsp->data+1);
 
-    if ((time < now.tv_sec) && ipmi_option_set_sel_time(mc->domain)) {
+    if ((time < (uint32_t)now.tv_sec) && ipmi_option_set_sel_time(mc->domain)) {
 	/* Time is in the past and setting time is requested, move it
 	   forward. */
 	DEBUG_INFO(mc->sel_timer_info);
